requirements.txt
pandas== 2.2.2
numpy==1.26.4
fastapi==0.112.1
uvicorn==0.30.6
python-multipart
pillow==10.4.0
tensorflow-serving-api==2.14.0
jinja2==3.1.2
requests==2.31.0
gunicorn==22.0.0
Flask==3.0.3
tensorflow
